Your commit message here

+++ Entities/Product.cs
@@ -5,7 +5,7 @@ namespace bakery.api.Entities
         public int ProductId { get; set; }
         public string ItemNumber { get; set; }
         public string ProductName { get; set; }
-        public float Price_Kg { get; set; }
+        public decimal Price_Kg { get; set; }
 
         public IList<SupplierProduct> SupplierProducts { get; set; }
